backend/requirements.txt
Django==2.2.12
djangorestframework==3.10.3
mysqlclient==1.4.6
djangorestframework-jwt==1.11.0
django-cors-headers==3.1.1
django-filter==2.2.0
gunicorn==19.9.0
django-rest-swagger==2.2.0
xhtml2pdf==0.2.3
numpy==1.17.2
pandas==0.25.1
channels==2.4.0
django-extensions==2.2.9
zeep==3.4.0
